Renamed publication_forms app to publication_workflow, as some of its new content will not be part of the publication forms. "publication_form.js" is still used to refer to the publication-form part of the publication workflow.
