+++ .gitignore
@@ -38,7 +38,7 @@ tardis/apps/*
 !tardis/apps/filepicker
 !tardis/apps/mx_views
 !tardis/apps/oaipmh
-!tardis/apps/publication_forms
+!tardis/apps/publication_workflow
 !tardis/apps/push_to
 !tardis/apps/related_info
 !tardis/apps/slideshow_view
