+++ docs/requirements.txt
@@ -1,4 +1,4 @@
 -r ../requirements-base.txt
 -r ../requirements-test.txt
 
--r ../tardis/apps/publication_forms/requirements.txt
+-r ../tardis/apps/publication_workflow/requirements.txt
