+++ tardis/apps/publication_workflow/static/publication-form/mx-dataset-description-template.html
