+++ tardis/apps/publication_workflow/static/publication-form/mx-pdb-template.html
