+++ tardis/apps/publication_workflow/static/publication-form/spinner.gif
