+++ tardis/tardis_portal/templates/tardis_portal/javascript_libraries.html
@@ -42,8 +42,8 @@
 <link type="text/css" href="{% static 'ng-dialog/css/ngDialog-theme-plain.min.css' %}" rel="stylesheet" />
 <link type="text/css" href="{% static 'ng-dialog/css/ngDialog-theme-default.min.css' %}" rel="stylesheet" />
 
-{% if pub_form_enabled %}
-<link type="text/css" href="{% static 'publication-form/publication_form.css' %}" rel="stylesheet" />
+{% if pub_workflow_enabled %}
+<link type="text/css" href="{% static 'publication-form/publication_workflow.css' %}" rel="stylesheet" />
 {% endif %}
 
 <!-- Twitter Bootstrap: http://twitter.github.com/bootstrap/ -->
